refactor: extract ArgumentNavigation to own file
